# apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Mendoans",
    "description": "Daftar API Mendoans",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/karyawans/{user_id}": {
      "get": {
        "tags": ["Karyawan"],
        "summary": "Get Data Karyawan",
        "description": "Menampilkan data karyawan yang dipilih",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User_id karyawan yang akan ditampilkan datanya"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found (user_id tidak terdaftar)"
          }
        }
      },
      "put": {
        "tags": ["Karyawan"],
        "summary": "Update Data Karyawan",
        "description": "edit data karyawan yang dipilih",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User_id karyawan yang akan di ubah datanya"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "fullname": {
                    "type": "string"
                  },
                  "statuspernikahan": {
                    "type": "string"
                  },
                  "nik": {
                    "type": "string"
                  },
                  "identitas": {
                    "type": "string"
                  },
                  "divisi": {
                    "type": "string"
                  },
                  "tanggalmasuk": {
                    "type": "string"
                  },
                  "statuskaryawan": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "alamat": {
                    "type": "string"
                  },
                  "posisi": {
                    "type": "string"
                  },
                  "site": {
                    "type": "string"
                  },
                  "telegram": {
                    "type": "string"
                  },
                  "createdby": {
                    "type": "string"
                  },
                  "createdate": {
                    "type": "string"
                  },
                  "pendidikan": {
                    "type": "string"
                  },
                  "institusi": {
                    "type": "string"
                  },
                  "tempatlahir": {
                    "type": "string"
                  },
                  "tanggallahir": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "role_trello": {
                    "type": "string"
                  },
                  "nikkaryawan": {
                    "type": "string"
                  },
                  "jurusan": {
                    "type": "string"
                  },
                  "createddby": {
                    "type": "string"
                  },
                  "updateat": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request (user_id harus unique)"
          },
          "404": {
            "description": "Not Found (user_id tidak ditemukan)"
          }
        }
      },
      "delete": {
        "tags": ["Karyawan"],
        "summary": "Delete Data Karyawan",
        "description": "hapus data karyawan yang dipilih",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User_id karyawan yang akan di hapus datanya"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}